acti_multi_agent/research-agent/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
networkx>=3.2
python-dotenv>=1.0.0
bibtexparser>=1.4.0
pyyaml>=6.0
numpy>=1.26.0

# Visualization dependencies (V1-V7)
plotly>=5.18.0
pyvis>=0.3.2
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
umap-learn>=0.5.5

# New tool integrations
nomic>=3.0.0
